Restructure and simplify API directories
